refactor(graphql): DRY scalar types via FhirStringScalarType base, drop dead FieldResolver helpers, fix ISchema alias
- Extract FhirStringScalarType abstract base; FhirDate/DateTime/Instant/Time all delegate to it
  (eliminates 4x identical ScalarType boilerplate)
- Remove GetArrayProperty and GetObjectProperty from FieldResolver — unused after FhirTypeModule
  switched to direct resolver lambdas
- Fix ISchema alias: was FhirISchema in ExperimentalAutofacRegistration, now consistent ISchema
- Drop unused IConfiguration parameter from AddGraphQlServices — only options are needed
