+++ src/Application/Ignixa.Application/Features/Experimental/GraphQl/Schema/FhirInstantScalarType.cs
@@ -3,24 +3,7 @@
 // Licensed under the MIT License. See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-using HotChocolate.Language;
-using HotChocolate.Types;
-
 namespace Ignixa.Application.Features.Experimental.GraphQl.Schema;
 
-internal sealed class FhirInstantScalarType : ScalarType<string, StringValueNode>
-{
-    public FhirInstantScalarType() : base("FhirInstant")
-    {
-        Description = "FHIR instant scalar (ISO 8601 with mandatory timezone)";
-    }
-
-    public override IValueNode ParseResult(object? resultValue)
-        => resultValue is string s ? new StringValueNode(s) : NullValueNode.Default;
-
-    protected override string ParseLiteral(StringValueNode valueSyntax)
-        => valueSyntax.Value;
-
-    protected override StringValueNode ParseValue(string runtimeValue)
-        => new(runtimeValue);
-}
+internal sealed class FhirInstantScalarType()
+    : FhirStringScalarType("FhirInstant", "FHIR instant scalar (ISO 8601 with mandatory timezone)");
